==> 06.03_Missing Indicators Imputation.ipynb <==
## This notebook inputs the missing values using Missing Indicators

import pandas as pd
from datetime import datetime
import numpy as np
from sklearn.impute import MissingIndicator

#### Choose wether the imputation should be done on the data set with daily or weekly granularity


INPUT, OUTPUT = 'output/CompleteWeeklyIndexes', 'output/MIIWeeklyImputed'

df = pd.read_csv(INPUT+'.csv', parse_dates=True, index_col=0)
df_test = df.loc['2015-01-02':] #Selecting test data
df = df.loc['2000-01-01':'2015-01-01'] #Selecting training data

indicator = MissingIndicator(missing_values=np.nan, features="all")

indicator.fit(df)
mask_all = indicator.transform(df)
mask_test_all = indicator.transform(df_test)

df_masks = pd.DataFrame(data=mask_all, index=df.index, columns=df.columns)
df_test_masks = pd.DataFrame(data=mask_test_all, index=df_test.index, columns=df_test.columns)

df_masks = df_masks.add_prefix('MI_')
df_test_masks = df_test_masks.add_prefix('MI_')

df_final = pd.concat([df, df_masks], axis=1, join='outer')
df_test_final = pd.concat([df_test, df_test_masks], axis=1, join='outer')

df_final.head()

df.head()

df_final=df_final.fillna(0)
df_test_final=df_test_final.fillna(0)
df_final.head()

df_final.to_csv(OUTPUT+'_training.csv', index = True)
df_test_final.to_csv(OUTPUT+ '_test.csv', index = True)
